Added new refset interface.

+++ ace-api/src/main/java/org/dwfa/ace/api/ebr/I_ThinExtByRefPartConceptLong.java
@@ -5,6 +5,6 @@ public interface I_ThinExtByRefPartConceptLong extends
 
 	   public long getLongValue();
 
-	    public void setIntValue(int longValue);
+	    public void setLongValue(long longValue);
 
 }
